rule_environment_view: use dash pages to create separate pages, introduce create_id to duplicate ids, create page to view query results by rule
