+++ frontend/config.py
@@ -1,0 +1,5 @@
+import os
+
+backend_root = os.getenv("MATCHER_BACKEND_URL")
+external_frontend_root = os.getenv("EXTERNAL_FRONTEND_URL")
+external_backend_root = os.getenv("EXTERNAL_BACKEND_URL")
