+++ frontend/dash_app.py
@@ -1,0 +1,21 @@
+from dash import Dash
+
+from frontend_api import app as Flask_app
+
+# imports callbacks required to register callbacks with the app
+import pages.rule.callbacks as rule_callbacks  # noqa
+import pages.snap.callbacks as snap_callbacks  # noqa
+
+
+app_dash = Dash(
+    __name__,
+    assets_folder="css/dash/",
+    routes_pathname_prefix='/dash/',
+    prevent_initial_callbacks=True,
+    server=Flask_app,
+    serve_locally=True,
+    use_pages=True,
+    suppress_callback_exceptions=True
+)
+
+server = app_dash.server
