+++ frontend/monkey.py
@@ -1,5 +1,4 @@
 from gevent import monkey
 monkey.patch_all()
 
-#from app import app
-from pairplot import server
+from dash_app import server  # noqa
