+++ frontend/pages/snap/constants.py
@@ -1,0 +1,5 @@
+DASH_PAGE_PREFIX = 'snap'
+
+
+def create_id(name: str) -> str:
+    return name
